+++ magda-web-client/src/Components/Account/AccountLoginPage.tsx
@@ -1,46 +1,20 @@
 import React, { useState, FunctionComponent } from "react";
-import fbLogo from "assets/login/fb-logo.svg";
-import googleLogo from "assets/login/google-logo.svg";
-import arcgisLogo from "assets/login/esri-logo.svg";
-import aafLogo from "assets/login/aaf-logo.png";
-import ckanLogo from "assets/login/ckan.png";
-import magdaLogo from "assets/login/magda.png";
-import genericLogo from "assets/login/generic-logo.svg";
 import "./AccountLoginPage.scss";
-import {
-    convertAuthPluginApiUrl,
-    getAuthProviders,
-    getAuthPlugins,
-    AuthConfig
-} from "api-clients/AuthApis";
+import { convertAuthPluginApiUrl, getAuthPlugins } from "api-clients/AuthApis";
+import { AuthPluginConfig } from "@magda/gateway/src/createAuthPluginRouter";
 import { config, isBackendSameOrigin } from "config";
 import { useAsync } from "react-async-hook";
-import urijs from "urijs";
 import markdownToHtml from "@magda/typescript-common/dist/markdownToHtml";
 import QrCodeLoginArea from "./QrCodeLoginArea";
+import { withRouter } from "react-router-dom";
+
 const { baseUrl, baseExternalUrl, authPluginRedirectUrl } = config;
 
 function getDefaultLoginFormProvider(
-    authConfigItems: AuthConfig[]
-): AuthConfig | null {
-    let selectConfigItem = authConfigItems.find(
-        (config) => !config.isAuthPlugin && config.config === "internal"
-    );
-    if (selectConfigItem) {
-        return selectConfigItem;
-    }
-
-    selectConfigItem = authConfigItems.find(
-        (config) => !config.isAuthPlugin && config.config === "ckan"
-    );
-    if (selectConfigItem) {
-        return selectConfigItem;
-    }
-
-    selectConfigItem = authConfigItems.find(
-        (config) =>
-            config.isAuthPlugin &&
-            config.config.authenticationMethod === "PASSWORD"
+    authConfigItems: AuthPluginConfig[]
+): AuthPluginConfig | null {
+    const selectConfigItem = authConfigItems.find(
+        (config) => config.authenticationMethod === "PASSWORD"
     );
     if (selectConfigItem) {
         return selectConfigItem;
@@ -49,96 +23,58 @@ function getDefaultLoginFormProvider(
     return null;
 }
 
-const makeLoginUrl = (authConfig: AuthConfig) => {
-    if (!authConfig.isAuthPlugin) {
-        const oauthRedirect = urijs()
-            .search("")
-            .fragment("")
-            .segment([])
-            .segment(`${baseUrl}sign-in-redirect`)
-            .search({
-                redirectTo: `${baseExternalUrl}account`
-            })
-            .toString();
-        return `${baseUrl}auth/login/${
-            authConfig.config
-        }?redirect=${encodeURIComponent(oauthRedirect)}`;
+const makeLoginUrl = (authConfig: AuthPluginConfig) => {
+    // auth plugins share the same gateway helm chart config [authPluginRedirectUrl](https://github.com/magda-io/magda/tree/master/deploy/helm/internal-charts/gateway)
+    // We don't need to supply redirection url here unless
+    // - the UI is served from different origin (domains) from backend.
+    // - And the configured redirect URL is not an absolute url
+    if (
+        isBackendSameOrigin ||
+        authPluginRedirectUrl.toLowerCase().indexOf("http") === 0
+    ) {
+        return `${baseUrl}auth/login/plugin/${authConfig.key}`;
     } else {
-        // auth plugins share the same gateway helm chart config [authPluginRedirectUrl](https://github.com/magda-io/magda/tree/master/deploy/helm/internal-charts/gateway)
-        // We don't need to supply redirection url here unless
-        // - the UI is served from different origin (domains) from backend.
-        // - And the configured redirect URL is not an absolute url
-        if (
-            isBackendSameOrigin ||
-            authPluginRedirectUrl.toLowerCase().indexOf("http") === 0
-        ) {
-            return `${baseUrl}auth/login/plugin/${authConfig.config.key}`;
-        } else {
-            return `${baseUrl}auth/login/plugin/${
-                authConfig.config.key
-            }?redirect=${encodeURIComponent(
-                baseExternalUrl +
-                    (authPluginRedirectUrl[0] === "/"
-                        ? authPluginRedirectUrl.substr(1)
-                        : authPluginRedirectUrl)
-            )}`;
-        }
+        return `${baseUrl}auth/login/plugin/${
+            authConfig.key
+        }?redirect=${encodeURIComponent(
+            baseExternalUrl +
+                (authPluginRedirectUrl[0] === "/"
+                    ? authPluginRedirectUrl.substr(1)
+                    : authPluginRedirectUrl)
+        )}`;
     }
 };
 
-function ucwords(str: string) {
-    str = str.toLowerCase();
-    return str.replace(/(^([a-zA-Z\p{M}]))|([ -][a-zA-Z\p{M}])/g, function (s) {
-        return s.toUpperCase();
-    });
-}
-
 type LoginFormPropsType = {
-    authConfig: AuthConfig;
+    authConfig: AuthPluginConfig;
 };
 
 const LoginFormArea: FunctionComponent<LoginFormPropsType> = (props) => {
     const { authConfig } = props;
 
-    let providerName: string = "";
-    let usernameLabel: string = "User name";
-    let passwordLabel: string = "Password";
+    let providerName = "";
+    let usernameLabel = "User name";
+    let passwordLabel = "Password";
     let loginFormExtraInfoHeading = "";
     let loginFormExtraInfoContent = "";
 
-    if (!authConfig.isAuthPlugin && authConfig.config === "internal") {
-        providerName = "Magda";
-        usernameLabel = "Email Address";
-        loginFormExtraInfoHeading = "Forgot your password?";
-        loginFormExtraInfoContent = `Forgot your password? ${
-            config.defaultContactEmail
-                ? `Email [${config.defaultContactEmail}](${config.defaultContactEmail})`
-                : "Contact your administrator."
-        }`;
-    } else if (!authConfig.isAuthPlugin && authConfig.config === "ckan") {
-        providerName = "Data.gov.au";
-        loginFormExtraInfoHeading = "Register";
-        loginFormExtraInfoContent = `To register a new data.gov.au account, [click here](https://data.gov.au/user/register)`;
-    } else if (
-        authConfig.isAuthPlugin &&
-        authConfig.config.authenticationMethod === "PASSWORD"
-    ) {
-        providerName = authConfig.config.name;
+    if (authConfig.authenticationMethod === "PASSWORD") {
+        providerName = authConfig.name;
 
-        usernameLabel = authConfig.config.loginFormUsernameFieldLabel
-            ? authConfig.config.loginFormUsernameFieldLabel
+        usernameLabel = authConfig.loginFormUsernameFieldLabel
+            ? authConfig.loginFormUsernameFieldLabel
             : usernameLabel;
 
-        passwordLabel = authConfig.config.loginFormPasswordFieldLabel
-            ? authConfig.config.loginFormPasswordFieldLabel
+        passwordLabel = authConfig.loginFormPasswordFieldLabel
+            ? authConfig.loginFormPasswordFieldLabel
             : passwordLabel;
 
-        loginFormExtraInfoHeading = authConfig.config.loginFormExtraInfoHeading
-            ? authConfig.config.loginFormExtraInfoHeading
+        loginFormExtraInfoHeading = authConfig.loginFormExtraInfoHeading
+            ? authConfig.loginFormExtraInfoHeading
             : loginFormExtraInfoHeading;
 
-        loginFormExtraInfoContent = authConfig.config.loginFormExtraInfoContent
-            ? authConfig.config.loginFormExtraInfoContent
+        loginFormExtraInfoContent = authConfig.loginFormExtraInfoContent
+            ? authConfig.loginFormExtraInfoContent
             : loginFormExtraInfoContent;
     } else {
         return null;
@@ -203,81 +139,46 @@ const LoginFormArea: FunctionComponent<LoginFormPropsType> = (props) => {
 };
 
 type LoginOptionList = {
-    authConfigItems: AuthConfig[] | undefined;
+    authConfigItems: AuthPluginConfig[] | undefined;
 };
 
-const isPasswordAuthItem = (config: AuthConfig): boolean => {
-    if (
-        !config.isAuthPlugin &&
-        (config.config === "internal" || config.config === "ckan")
-    ) {
-        return true;
-    } else if (
-        config.isAuthPlugin &&
-        config.config.authenticationMethod === "PASSWORD"
-    ) {
+const isPasswordAuthItem = (config: AuthPluginConfig): boolean => {
+    if (config.authenticationMethod === "PASSWORD") {
         return true;
     } else {
         return false;
     }
 };
 
-const getSortingStringFromAuthConfig = (config: AuthConfig): string =>
-    config.isAuthPlugin ? config.config.name : config.config;
+const getSortingStringFromAuthConfig = (config: AuthPluginConfig): string =>
+    config.name;
 
-export default function Login(props) {
+function AccountLoginPage(props) {
     const {
         result: authConfigItems,
-        loading: isProvidersLoading,
-        error: providersLoadingError
+        loading: isAuthPluginsLoading,
+        error: authPluginsLoadingError
     } = useAsync(async () => {
-        let commonAuthModuleConfig: AuthConfig[] = [];
-        let authPlugins;
-        let providers;
+        let authPlugins: AuthPluginConfig[] = [];
 
         try {
-            authPlugins = (await getAuthPlugins()).map(
-                (item) =>
-                    ({
-                        isAuthPlugin: true,
-                        config: item
-                    } as AuthConfig)
-            );
-            commonAuthModuleConfig = [
-                ...commonAuthModuleConfig,
-                ...authPlugins
-            ];
+            authPlugins = await getAuthPlugins();
         } catch (e) {
             console.error(e);
         }
 
-        try {
-            providers = (await getAuthProviders()).map(
-                (item) =>
-                    ({
-                        isAuthPlugin: false,
-                        config: item
-                    } as AuthConfig)
-            );
-            commonAuthModuleConfig = [...commonAuthModuleConfig, ...providers];
-        } catch (e) {
-            if (typeof authPlugins === "undefined") {
-                // -- if both authPlugins & provider loading failed
-                throw e;
-            } else {
-                console.error(e);
-            }
-        }
-
         let defaultSelectedAuthConfig = getDefaultLoginFormProvider(
-            commonAuthModuleConfig
+            authPlugins
         );
 
-        if (!defaultSelectedAuthConfig && commonAuthModuleConfig.length) {
-            defaultSelectedAuthConfig = commonAuthModuleConfig[0];
+        if (!defaultSelectedAuthConfig && authPlugins.length) {
+            defaultSelectedAuthConfig = authPlugins[0];
         }
 
-        function sortAuthConfigItems(a: AuthConfig, b: AuthConfig): number {
+        function sortAuthConfigItems(
+            a: AuthPluginConfig,
+            b: AuthPluginConfig
+        ): number {
             if (defaultSelectedAuthConfig === a) {
                 // --- selected default provider / plugin always show as no.1
                 return -1;
@@ -295,17 +196,17 @@ export default function Login(props) {
             }
         }
 
-        commonAuthModuleConfig.sort(sortAuthConfigItems);
+        authPlugins.sort(sortAuthConfigItems);
 
         setSelectedAuthConfig(defaultSelectedAuthConfig);
 
-        return commonAuthModuleConfig;
+        return authPlugins;
     }, []);
 
     const [
         selectedAuthConfig,
         setSelectedAuthConfig
-    ] = useState<AuthConfig | null>(null);
+    ] = useState<AuthPluginConfig | null>(null);
 
     const LoginOptionList: FunctionComponent<LoginOptionList> = (props) => {
         const { authConfigItems } = props;
@@ -317,135 +218,63 @@ export default function Login(props) {
         return (
             <ul className="login__providers">
                 {authConfigItems.map((item, idx) => {
-                    if (!item.isAuthPlugin && item.config === "internal") {
+                    if (item.authenticationMethod === "PASSWORD") {
                         return (
                             <li key={idx} className="login__provider">
                                 <a onClick={() => setSelectedAuthConfig(item)}>
                                     <img
-                                        src={magdaLogo}
+                                        src={convertAuthPluginApiUrl(
+                                            item.key,
+                                            item.iconUrl
+                                        )}
                                         className="login__logo"
                                         alt="logo"
                                     />
-                                    Magda
+                                    {item.name}
                                 </a>
                             </li>
                         );
-                    } else if (!item.isAuthPlugin && item.config === "ckan") {
-                        return (
-                            <li key={idx} className="login__provider">
-                                <a onClick={() => setSelectedAuthConfig(item)}>
-                                    <img
-                                        src={ckanLogo}
-                                        className="login__logo"
-                                        alt="logo"
-                                    />
-                                    Data.gov.au / Ckan
-                                </a>
-                            </li>
-                        );
-                    } else if (!item.isAuthPlugin) {
-                        let logo = genericLogo;
-                        let name = ucwords(item.config);
-                        switch (item.config) {
-                            case "facebook":
-                                logo = fbLogo;
-                                break;
-                            case "google":
-                                logo = googleLogo;
-                                break;
-                            case "arcgis":
-                                logo = arcgisLogo;
-                                name = "Esri";
-                                break;
-                            case "aaf":
-                                logo = aafLogo;
-                                name = "AAF";
-                                break;
-                            // vanguard will use the default values
-                        }
-
+                    } else if (
+                        item.authenticationMethod === "IDP-URI-REDIRECTION"
+                    ) {
                         return (
                             <li key={idx} className="login__provider">
                                 <a href={makeLoginUrl(item)}>
                                     <img
-                                        src={logo}
+                                        src={convertAuthPluginApiUrl(
+                                            item.key,
+                                            item.iconUrl
+                                        )}
                                         className="login__logo"
                                         alt="logo"
                                     />
-                                    {name}
+                                    {item.name}
+                                </a>
+                            </li>
+                        );
+                    } else if (item.authenticationMethod === "QR-CODE") {
+                        return (
+                            <li key={idx} className="login__provider">
+                                <a onClick={() => setSelectedAuthConfig(item)}>
+                                    <img
+                                        src={convertAuthPluginApiUrl(
+                                            item.key,
+                                            item.iconUrl
+                                        )}
+                                        className="login__logo"
+                                        alt="logo"
+                                    />
+                                    {item.name}
                                 </a>
                             </li>
                         );
                     } else {
-                        if (item.config.authenticationMethod === "PASSWORD") {
-                            return (
-                                <li key={idx} className="login__provider">
-                                    <a
-                                        onClick={() =>
-                                            setSelectedAuthConfig(item)
-                                        }
-                                    >
-                                        <img
-                                            src={convertAuthPluginApiUrl(
-                                                item.config.key,
-                                                item.config.iconUrl
-                                            )}
-                                            className="login__logo"
-                                            alt="logo"
-                                        />
-                                        {item.config.name}
-                                    </a>
-                                </li>
-                            );
-                        } else if (
-                            item.config.authenticationMethod ===
-                            "IDP-URI-REDIRECTION"
-                        ) {
-                            return (
-                                <li key={idx} className="login__provider">
-                                    <a href={makeLoginUrl(item)}>
-                                        <img
-                                            src={convertAuthPluginApiUrl(
-                                                item.config.key,
-                                                item.config.iconUrl
-                                            )}
-                                            className="login__logo"
-                                            alt="logo"
-                                        />
-                                        {item.config.name}
-                                    </a>
-                                </li>
-                            );
-                        } else if (
-                            item.config.authenticationMethod === "QR-CODE"
-                        ) {
-                            return (
-                                <li key={idx} className="login__provider">
-                                    <a
-                                        onClick={() =>
-                                            setSelectedAuthConfig(item)
-                                        }
-                                    >
-                                        <img
-                                            src={convertAuthPluginApiUrl(
-                                                item.config.key,
-                                                item.config.iconUrl
-                                            )}
-                                            className="login__logo"
-                                            alt="logo"
-                                        />
-                                        {item.config.name}
-                                    </a>
-                                </li>
-                            );
-                        } else {
-                            return (
-                                <li
-                                    key={idx}
-                                    className="login__provider"
-                                >{`Unrecognised Auth Plugin: ${item.config.name}`}</li>
-                            );
-                        }
+                        return (
+                            <li
+                                key={idx}
+                                className="login__provider"
+                            >{`Unrecognised Auth Plugin: ${item.name}`}</li>
+                        );
                     }
                 })}
             </ul>
@@ -459,7 +288,7 @@ export default function Login(props) {
                     <p>Please select an authentication option from the list.</p>
                 );
             }
-            if (isProvidersLoading) {
+            if (isAuthPluginsLoading) {
                 return null;
             } else {
                 return (
@@ -475,13 +304,10 @@ export default function Login(props) {
                     <LoginFormArea authConfig={selectedAuthConfig} />
                 </>
             );
-        } else if (
-            selectedAuthConfig.isAuthPlugin &&
-            selectedAuthConfig.config.authenticationMethod === "QR-CODE"
-        ) {
+        } else if (selectedAuthConfig.authenticationMethod === "QR-CODE") {
             return (
                 <>
-                    <QrCodeLoginArea authConfig={selectedAuthConfig.config} />
+                    <QrCodeLoginArea authConfig={selectedAuthConfig} />
                 </>
             );
         } else {
@@ -491,29 +317,29 @@ export default function Login(props) {
 
     return (
         <div className="row login__row">
-            {props.signInError && (
+            {props?.signInError && (
                 <div className="col-xs-12">
                     <div className="au-body au-page-alerts au-page-alerts--error">
-                        <p>Sign In Failed: {props.signInError} </p>
+                        <p>Error: {props.signInError} </p>
                     </div>
                 </div>
             )}
-            {isProvidersLoading ? (
+            {isAuthPluginsLoading ? (
                 <div className="col-xs-12">
-                    <p>Loading available authentication providers...</p>
+                    <p>Loading available authentication options...</p>
                 </div>
             ) : null}
-            {!isProvidersLoading && providersLoadingError ? (
+            {!isAuthPluginsLoading && authPluginsLoadingError ? (
                 <div className="col-xs-12">
                     <div className="au-body au-page-alerts au-page-alerts--error">
                         <p>
-                            Failed to load authentication providers:{" "}
-                            {"" + providersLoadingError}{" "}
+                            Failed to load authentication plugin:{" "}
+                            {"" + authPluginsLoadingError}{" "}
                         </p>
                     </div>
                 </div>
             ) : null}
-            {!isProvidersLoading ? (
+            {!isAuthPluginsLoading ? (
                 <div className="col-sm-6 col-md-5">
                     <h2>Sign In / Register Providers</h2>
                     <LoginOptionList authConfigItems={authConfigItems} />
@@ -524,3 +350,5 @@ export default function Login(props) {
         </div>
     );
 }
+
+export default withRouter(AccountLoginPage);