+++ magda-web-client/src/api-clients/AuthApis.ts
@@ -12,16 +12,6 @@ import {
     ResourceRecord
 } from "@magda/typescript-common/dist/authorization-api/model";
 
-export type AuthConfig =
-    | {
-          isAuthPlugin: false;
-          config: string;
-      }
-    | {
-          isAuthPlugin: true;
-          config: AuthPluginConfig;
-      };
-
 export type QrCodeImgDataResponse = {
     token: string;
     data: string;
@@ -31,18 +21,6 @@ export type QrCodePollResponse = {
     result: "pending" | "success" | "failure";
     errorMessage: string;
 };
-
-export async function getAuthProviders(): Promise<string[]> {
-    const providers = await request(
-        "GET",
-        getAbsoluteUrl(`auth/providers`, config.baseUrl)
-    );
-    if (providers) {
-        return providers;
-    } else {
-        return [];
-    }
-}
 
 export async function getAuthPlugins(): Promise<AuthPluginConfig[]> {
     const plugins = await request(