+++ CHANGES.md
@@ -29,6 +29,14 @@
 - #3345 Add dataset index / delete API endpoints to indexer
 - Add `debug` switch to registry API, search API & Auth API
 
+## 1.3.0
+
+- Upgrade nodemailer to 6.7.3
+- Upgrade urijs to 1.19.11 and @types/urijs to 1.19.19
+- #3020 Remove all legacy auth provider from gateway and convert them into auth plugins in separate repos
+- Add `authPluginAllowedExternalRedirectDomains` support to auth plugin spec & rewrite relevant packages
+- Redirect users to login page with proper error message, when they try to access the admin pages that they don't have permissions.
+
 ## 1.2.1
 
 - Upgrade magda-csw-connector to v1.1.1 for [license extraction issue on certain sources](https://github.com/magda-io/magda-csw-connector/issues/21)